project inital setup

+++ index.js
@@ -3,6 +3,7 @@ let program = require('commander');
 let shell = require('shelljs');
 let colors = require('colors');
 let fs = require('fs-extra');
+const inquirer = require('inquirer');
 const replace = require('replace');
 const template = require('./component_template/template');
 let appName;
@@ -10,20 +11,66 @@ let appDirectory;
 let newCompPath;
 let nofolder;
 let functional;
-let observable;
 let stylesheet;
+
 program
-  .version('2.0.8')
-  .command('init <dir>')
+  .version('1.0.0')
+  .command('create <dir>')
   .option('-T , --typscript', 'Install with typescript')
-  .action(createReact);
+  .action((dir) => {
+    let project;
+    var questions = [{
+      type: 'list',
+      name: 'project',
+      message: "Which project you want to create?",
+      choices: ['ANGULAR', 'REACT', 'REACT NATIVE']
+    }]
+
+    inquirer.prompt(questions).then(answers => {
+      project = answers['project']
+      switch (project) {
+        case 'REACT':
+          createReact(dir)
+          break;
+        case 'REACT NATIVE':
+          console.log('working on react native'.green);
+          break;
+        case 'ANGULAR':
+          console.log('working on angular'.green);
+          break;
+  
+  
+        default:
+          console.log(`working on ${project} new feature`.green);
+          break;
+      }
+    })
+  });
+
+program
+  .command('test')
+  .action(() => {
+
+  })
+
 program
   .command('gc <component>')
   .option('-n, --nofolder', 'Do not wrap component in folder')
-  .option('-o, --observable', 'Make observable')
-  .option('-s, --style', 'With stylesheet')
+  .option('-ws, --nostyle', 'Without stylesheet', true)
   .option('-f, --functional', 'Create functional component')
   .action(createComponent);
+
+program
+  .command('generate-component <component>')
+  .option('-n, --nofolder', 'Do not wrap component in folder')
+  .option('-ws, --nostyle', 'Without stylesheet')
+  .option('-f, --functional', 'Create functional component')
+  .action(createComponent);
+
+program
+  .command('run <cmd>')
+  .action(npmCommandRunner)
+
 program.parse(process.argv)
 async function createReact(dir) {
   appName = dir;
@@ -40,17 +87,27 @@ async function createReact(dir) {
     await installPackages();
     await updatePackage_json();
     await generateBoilerplate();
-    shell.exec(`npm install`, { cwd: appDirectory }, (e) => {
-      console.log("All done");
+    shell.exec(`npm install`, { cwd: appDirectory }, (e, stdout, stderr) => {
+      if (stderr) {
+        console.log(stdout);
+        console.log("All done".green);
+      }
     });
 
   }
 
 }
+function npmCommandRunner(cmd = 'start') {
+  shell.exec(`npm run ${cmd}`, (e, stdout, stderr) => {
+    if (e.toString().search('Something is already running on port 3000.')) {
+      console.log(`Error ⇑`.red);
+    }
+  })
+}
 function createReactApp() {
   return new Promise(resolve => {
     if (appName) {
-      console.log("\nCreating react app...".cyan);
+      console.log("\nCreating react app...".green);
       try {
         shell.exec(`node ${require('path').dirname(require.main.filename)}/node_modules/create-react-app/index.js ${appName}`, (e, stdout, stderr) => {
           if (stderr) {
@@ -68,7 +125,7 @@ function createReactApp() {
         });
       } catch (e) {
         console.log('create-react-app not installed'.red);
-        console.log("\nInstalling create react app...".cyan);
+        console.log("\nInstalling create react app...".green);
         shell.exec(`npm install -g create-react-app`, (e, stdout, stderr) => {
           console.log("Finished installing creating react app".green);
           createReactApp();
@@ -80,7 +137,7 @@ function createReactApp() {
     } else {
       console.log("\nNo app name was provided.".red);
       console.log("\nProvide an app name in the following format: ");
-      console.log("\ncreate-nick-react ", "app-name\n".cyan);
+      console.log("\ncreate-nick-react ", "app-name\n".green);
       resolve(false);
       process.exit(1);
     }
@@ -88,8 +145,8 @@ function createReactApp() {
 }
 function installPackages() {
   return new Promise(resolve => {
-    console.log("\nInstalling react-router, react-router-dom, react-lazy-route, axios, sass and redux...".cyan);
-    shell.exec(`npm install --save react-router react-router-dom react-lazy-route axios mobx mobx-react node-sass-chokidar npm-run-all`, { cwd: appDirectory }, (e) => {
+    console.log("\nInstalling react-router, react-router-dom, axios, redux-thunk and redux...".green);
+    shell.exec(`npm install --save react-router react-router-dom axios redux react-redux redux-thunk`, { cwd: appDirectory }, (e) => {
       console.log("\nFinished installing packages\n".green);
       resolve()
     });
@@ -97,17 +154,13 @@ function installPackages() {
 }
 function updatePackage_json() {
   let scripts = {
-    "build-css": "node-sass-chokidar src/ -o src/",
-    "watch-css": "npm run build-css && node-sass-chokidar src/ -o src/ --watch --recursive",
-    "start-js": "react-scripts start",
-    "start": "npm-run-all -p watch-css start-js",
-    "build-js": "react-scripts build",
-    "build": "npm-run-all build-css build-js",
-    "test": "react-scripts test --env=jsdom",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "echo \"Error: no test specified\" && exit 1",
     "eject": "react-scripts eject"
   };
   return new Promise(resolve => {
-    console.log("\nUpdating package.json....".cyan);
+    console.log("\nUpdating package.json....".green);
     shell.exec(`node ${require('path').dirname(require.main.filename)}/node_modules/json/lib/json.js -I -f package.json -e 'this.scripts=` + JSON.stringify(scripts) + "'", { cwd: appDirectory }, () => {
       resolve();
     });
@@ -115,14 +168,17 @@ function updatePackage_json() {
 }
 function generateBoilerplate() {
   console.log();
-  console.log("\nGenerating boilerplate...".cyan);
+  console.log("\nGenerating boilerplate...".green);
   return new Promise(resolve => {
-    fs.unlinkSync(`${appDirectory}/src/App.css`);
-    fs.unlinkSync(`${appDirectory}/src/App.js`);
-    fs.unlinkSync(`${appDirectory}/src/App.test.js`);
-    fs.unlinkSync(`${appDirectory}/src/index.css`);
-    fs.unlinkSync(`${appDirectory}/src/logo.svg`);
-    fs.copySync(`${require('path').dirname(require.main.filename)}/templates`, `${appDirectory}/src`);
+    if (fs.existsSync(`${appDirectory}/src/App.js`)) {
+      fs.unlinkSync(`${appDirectory}/src/App.css`);
+      fs.unlinkSync(`${appDirectory}/src/App.js`);
+      fs.unlinkSync(`${appDirectory}/src/App.test.js`);
+      fs.unlinkSync(`${appDirectory}/src/index.css`);
+      fs.unlinkSync(`${appDirectory}/src/logo.svg`);
+    }
+    fs.copySync(`${require('path').dirname(require.main.filename)}/templates/react-redux/src`, `${appDirectory}/src`);
+    fs.copySync(`${require('path').dirname(require.main.filename)}/templates/react-redux/public`, `${appDirectory}/public`);
     resolve();
   })
 }
@@ -131,10 +187,12 @@ async function createComponent(component, cmd) {
   cmd.nofolder ? nofolder = true : nofolder = false;
   cmd.functional ? functional = true : functional = false;
   cmd.observable ? observable = true : observable = false;
-  cmd.style ? stylesheet = true : stylesheet = false;
-  if (fs.existsSync('./src/components')) {
-    newCompPath = `./src/components/${component}`;
-  }
+  cmd.nostyle ? stylesheet = false : stylesheet = true;
+  // if (fs.existsSync('./src/components')) {
+  //   newCompPath = `./src/components/${component}`;
+  // } else {
+  //   component = `./src/components/${component}`;
+  // }
   let template = await buildTemplate();
   writeFile(template, component)
 }
@@ -151,7 +209,12 @@ function buildTemplate() {
   return imports.join('\n') + '\n' + body + '\n' + exported;
 }
 function capitalize(comp) {
-  return comp[0].toUpperCase() + comp.substring(1, comp.length);
+  const words = comp.split('-');
+  let compName = "";
+  words.forEach(word => {
+    compName += word[0].toUpperCase() + word.substring(1, word.length);
+  })
+  return compName;
 }
 function writeFile(template, component) {
   let path = newCompPath;
@@ -172,7 +235,7 @@ function writeFile(template, component) {
     if (!fs.existsSync(`${path}.scss`)) {
       console.log('creating syles');
       fs.outputFileSync(`${path}.scss`, '');
-      console.log(`Stylesheet ${comp} created at ${path}.scss`.cyan)
+      console.log(`Stylesheet ${comp} created at ${path}.scss`.green)
     } else {
       console.log(`Stylesheet ${comp} allready exists at ${path}.scss, choose another name if you want to create a new stylesheet`.red)
     }
@@ -187,7 +250,7 @@ function writeFile(template, component) {
         recursive: false,
         silent: true,
       });
-      console.log(`Component ${comp} created at ${path}.js`.cyan)
+      console.log(`Component ${comp} created at ${path}.js`.green)
     });
   } else {
     console.log(`Component ${comp} allready exists at ${path}.js, choose another name if you want to create a new component`.red)
